pyproject.toml
[project]
name = "agent-service-toolkit"
version = "0.1.0"
description = "Full toolkit for running an AI agent service built with LangGraph, FastAPI and Streamlit"
readme = "README.md"
authors = [{ name = "Joshua Carroll", email = "carroll.joshk@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.11"

dependencies = [
    "duckduckgo-search==8.0.1",
    "fastapi==0.115.12",
    "grpcio==1.71.0",
    "httpx==0.27.2",
    "jiter==0.8.2",
    "langchain-core==0.3.59",
    "langchain-community==0.3.23",
    "langchain-anthropic==0.3.13",
    "langchain-aws==0.2.23",
    "langchain-chroma==0.2.3",
    "langchain-google-genai==2.0.11",
    "langchain-google-vertexai==2.0.7",
    "langchain-groq==0.2.5",
    "langchain-ollama==0.2.3",
    "langchain-openai==0.3.16",
    "langgraph==0.3.34",
    "langgraph-checkpoint-mongodb==0.1.3",
    "langgraph-checkpoint-postgres==2.0.21",
    "langgraph-checkpoint-sqlite==2.0.7",
    "langgraph-supervisor==0.0.21",
    "langsmith==0.1.147",
    "numexpr==2.10.2",
    "numpy ~=1.26.4; python_version <= '3.12'",
    "numpy ~=2.2.3; python_version >= '3.13'",
    "onnxruntime==1.22.0",
    "pandas==2.2.3",
    "psycopg[binary,pool]==3.2.7",
    "pyarrow==20.0.0",
    "pydantic==2.10.6",
    "pydantic-settings==2.6.1",
    "pyowm==3.3.0",
    "python-dotenv==1.0.1",
    "setuptools==75.6.0",
    "streamlit==1.40.2",
    "tiktoken==0.9.0",
    "uvicorn==0.32.1",
    "pypdf==5.3.1",
    "docx2txt==0.9",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.51.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "boto3==1.38.13",
    "botocore==1.38.13",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.2",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "click==8.2.0",
    "coloredlogs==15.0.1",
    "comm==0.2.2",
    "coverage==7.8.0",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "deprecated==1.2.18",
    "distlib==0.3.9",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docstring-parser==0.16",
    "durationpy==0.9",
    "executing==2.2.0",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "gdown==5.2.0",
    "geojson==2.5.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.24.2",
    "google-auth==2.40.1",
    "google-cloud-aiplatform==1.92.0",
    "google-cloud-bigquery==3.31.0",
    "google-cloud-core==2.4.3",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.7.1",
    "google-genai==1.2.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.2",
    "groq==0.24.0",
    "grpc-google-iam-v1==0.14.2",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "hf-xet==1.1.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.31.1",
    "humanfriendly==10.0",
    "identify==2.6.10",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "iniconfig==2.1.0",
    "ipykernel==6.29.5",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "kubernetes==32.0.1",
    "langchain==0.3.25",
    "langchain-tavily==0.2.1",
    "langchain-text-splitters==0.3.8",
    "langfuse==2.60.5",
    "langgraph-checkpoint==2.0.25",
    "langgraph-prebuilt==0.1.8",
    "langgraph-sdk==0.1.66",
    "lxml==5.4.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "motor==3.7.0",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "mypy==1.15.0",
    "mypy-extensions==1.1.0",
    "narwhals==1.38.2",
    "nest-asyncio==1.6.0",
    "nodeenv==1.9.1",
    "oauthlib==3.2.2",
    "ollama==0.4.8",
    "openai==1.78.0",
    "opentelemetry-api==1.33.0",
    "opentelemetry-exporter-otlp-proto-common==1.33.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.33.0",
    "opentelemetry-instrumentation==0.54b0",
    "opentelemetry-instrumentation-asgi==0.54b0",
    "opentelemetry-instrumentation-fastapi==0.54b0",
    "opentelemetry-proto==1.33.0",
    "opentelemetry-sdk==1.33.0",
    "opentelemetry-semantic-conventions==0.54b0",
    "opentelemetry-util-http==0.54b0",
    "orjson==3.10.18",
    "ormsgpack==1.9.1",
    "overrides==7.7.0",
    "packaging==24.2",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "pluggy==1.5.0",
    "posthog==4.0.1",
    "pre-commit==4.2.0",
    "primp==0.15.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "psycopg-binary==3.2.7",
    "psycopg-pool==3.2.6",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic-core==2.27.2",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "pymongo==4.11.3",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pysocks==1.7.1",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "pytest-cov==6.1.1",
    "pytest-env==1.1.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "ruff==0.11.9",
    "s3transfer==0.12.0",
    "shapely==2.1.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.40",
    "stack-data==0.6.3",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tokenizers==0.21.1",
    "toml==0.10.2",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.15.3",
    "typing-inspect==0.9.0",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvloop==0.21.0",
    "virtualenv==20.31.2",
    "watchdog==6.0.0",
    "watchfiles==1.0.5",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "wrapt==1.17.2",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zipp==3.21.0",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

# Group for the minimal dependencies to run just the client and Streamlit app.
# These are also installed in the default dependencies.
# To install run: `uv sync --frozen --only-group client`
client = [
    "httpx~=0.27.2",
    "pydantic ~=2.10.1",
    "python-dotenv ~=1.0.1",
    "streamlit~=1.40.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
OPENAI_API_KEY = "sk-fake-openai-key"

[tool.mypy]
plugins = "pydantic.mypy"
exclude = "src/streamlit_app.py"

[[tool.mypy.overrides]]
module = ["numexpr.*"]
follow_untyped_imports = true
